loading levels works, manu operations, back from game...

+++ www/js/script.js
@@ -1,11 +1,27 @@
 // in create: place create instead of printPieces 
-var game = new Phaser.Game(480, 640, Phaser.CANVAS, null, {preload: preload, create: printPieces, update: update});
+var game = new Phaser.Game(480, 640, Phaser.CANVAS, null, {preload: preload, create: create, update: update});
 
-var playButton, exitButton, levels, prevButton, nextButton, textChoose, textLevel, startButton, backButton, level, maxLevel;
+var playButton, exitButton, levelPrev, level, prevButton, nextButton, textChoose, textLevel, startButton, backButton, level, maxLevel;
 
 piece = [];
-maxLevel = 5;
+level = 0;
+maxLevel = 2;
 level = maxLevel;
+
+pieceInfo = [
+    [
+        { "left" : 100, "top" : 250, "frame" : 1, "linked"  : [ 0 ]         },
+    ],
+    [
+        { "left" : 150, "top" : 250, "frame" : 1, "linked"  : [ 0 ]         },
+        { "left" : 250, "top" : 250, "frame" : 2, "linked"  : [ 0, 1 ]      }
+    ],
+    [
+        { "left" : 150, "top" : 250, "frame" : 1, "linked"  : [ 0 ]         },
+        { "left" : 270, "top" : 350, "frame" : 0, "linked"  : [ 0, 2 ]      },
+        { "left" : 320, "top" : 250, "frame" : 2, "linked"  : [ 1, 2 ]      }
+    ]
+]
 
 function preload() {
 
@@ -15,11 +31,12 @@ function preload() {
     game.stage.backgroundColor = '#eee';
     game.load.spritesheet('playBtn', 'img/play12.png', 150, 50);
     game.load.spritesheet('exitBtn', 'img/exit12.png', 150, 50);
-    game.load.spritesheet('levels', 'img/levels.png', 200, 200);
+    game.load.spritesheet('levelPrev', 'img/levels.png', 200, 200);
     game.load.spritesheet('prevBtn', 'img/prev.png', 70, 50);
     game.load.spritesheet('nextBtn', 'img/next.png', 70, 50);
     game.load.spritesheet('startBtn', 'img/start12.png', 150, 50);
     game.load.spritesheet('backBtn', 'img/back12.png', 150, 50);
+    game.load.spritesheet('gameBackBtn', 'img/back12.png', 150, 50);
     game.load.spritesheet('piece', 'img/piece.png', 50, 50);
 }
 function create() {
@@ -32,16 +49,17 @@ function create() {
 function update() {}
 
 function playGame(){
-    playButton.kill();
-    exitButton.kill();
+    if(typeof playButton !== 'undefined') playButton.kill();
+    if(typeof exitButton !== 'undefined') exitButton.kill();
+    if(typeof gameBackButton !== 'undefined') gameBackButton.kill();
 
-    levelImg = game.add.sprite(game.world.width*0.5, game.world.height*0.4, 'levels');
+    levelImg = game.add.sprite(game.world.width*0.5, game.world.height*0.4, 'levelPrev');
     levelImg.anchor.set(0.5);
-    levelImg.frame = maxLevel-1;
-	
+    levelImg.frame = level;
+    
     prevButton = game.add.button(game.world.width*0.3, game.world.height*0.65, 'prevBtn', prevGame, this, 0, 0, 1, 2);
     prevButton.anchor.set(0.5);
-	nextButton = game.add.button(game.world.width*0.7, game.world.height*0.65, 'nextBtn', nextGame, this, 0, 0, 1, 2);
+    nextButton = game.add.button(game.world.width*0.7, game.world.height*0.65, 'nextBtn', nextGame, this, 0, 0, 1, 2);
     nextButton.anchor.set(0.5);
     textChoose = game.add.text(game.world.width*0.5, game.world.height*0.2, "Choose a game level:", {
         font: "25px Arial",
@@ -50,7 +68,7 @@ function playGame(){
     });
     textChoose.anchor.setTo(0.5);
 
-    textLevel = game.add.text(game.world.width*0.5, game.world.height*0.65, level, {
+    textLevel = game.add.text(game.world.width*0.5, game.world.height*0.65, level+1, {
         font: "25px Arial",
         fill: "#ff0044",
         align: "center"
@@ -64,15 +82,15 @@ function playGame(){
 }
 function prevGame(){
     level--;
-    if (level < 1) level = 1;
-    levelImg.frame = level-1;
-    textLevel.setText(level);
+    if (level < 0) level = 0;
+    levelImg.frame = level;
+    textLevel.setText(level+1);
 }
 function nextGame(){
     level++;
     if (level > maxLevel) level = maxLevel; 
-    levelImg.frame = level-1;
-    textLevel.setText(level);
+    levelImg.frame = level;
+    textLevel.setText(level+1);
 }
 function startGame(){
     levelImg.kill();
@@ -84,26 +102,29 @@ function startGame(){
     backButton.kill();
     
     printPieces();
+
+    gameBackButton = game.add.button(game.world.width*0.5, game.world.height*0.8, 'gameBackBtn', gameMenu, this, 0, 0, 1, 2);
+    gameBackButton.anchor.set(0.5);
 }
-//function that shows pieces on page for that level
+function gameMenu(){
+    removePieces();
+    playGame();
+}
+//function that shows pieces on page for that levels
 function printPieces(){
-
-pieceInfo = [
-    [
-        { "left" : 150, "top" : 250, "frame" : 1, "linked"  : [ 0 ]         },
-        { "left" : 250, "top" : 250, "frame" : 2, "linked"  : [ 0, 1 ]      }
-    ]
-]
-level = 0;
-levelPiecesNum = pieceInfo[level].length;
-
+    levelPiecesNum = pieceInfo[level].length;
     for (var i = 0; i < levelPiecesNum; i++){
         piece[i] = game.add.button(pieceInfo[level][i].left, pieceInfo[level][i].top, 'piece', pieceClick.bind(this, i), this); 
         piece[i].frame = pieceInfo[level][i].frame;
         piece[i].anchor.set(0.5);
     }
 }
-
+function removePieces(){
+    levelPiecesNum = pieceInfo[level].length;
+    for (var i = 0; i < levelPiecesNum; i++){
+        piece[i].kill();
+    }
+}
 function backGame(){
     levelImg.kill();
     prevButton.kill();
@@ -121,15 +142,24 @@ function pieceClick(pieceNum){
         tempPieceNum = pieceInfo[level][pieceNum].linked[i];
         piece[tempPieceNum].frame = piece[tempPieceNum].frame + 1;
     };
-    isSolved();
+    isSolved(pieceNum);
 }
-function isSolved(){
+function isSolved(pieceNum){
     var solved = 1;
     for (var i = 0; i < levelPiecesNum; i++) {
         if(piece[i].frame != 0) solved = 0;
     };
     if (solved == 1){
         console.log("Solved");
+                
+        piece[pieceNum].inputEnabled = false;
+        
+        setTimeout(function(){ 
+            removePieces(); 
+            level++; 
+            printPieces() 
+        }, 500);
+
     } else {
         console.log("NOT solved");
     }
